move BlockHeight and BlockWeight into their own modules
