+++ src/Chainweb/BlockHeader.hs
@@ -42,20 +42,6 @@ module Chainweb.BlockHeader
 -- $guards
 , slowEpochGuard
 
--- * Block Height
-, BlockHeight(..)
-, encodeBlockHeight
-, decodeBlockHeight
-, encodeBlockHeightBe
-, decodeBlockHeightBe
-
--- * Block Weight
-, BlockWeight(..)
-, encodeBlockWeight
-, decodeBlockWeight
-, encodeBlockWeightBe
-, decodeBlockWeightBe
-
 -- * Block Payload Hash
 , BlockPayloadHash(..)
 , encodeBlockPayloadHash
@@ -165,6 +151,8 @@ import GHC.Generics (Generic)
 -- Internal imports
 
 import Chainweb.BlockHash
+import Chainweb.BlockHeight
+import Chainweb.BlockWeight
 import Chainweb.ChainId
 import Chainweb.Crypto.MerkleLog
 import Chainweb.Difficulty
@@ -180,7 +168,6 @@ import Chainweb.Version
 
 import Data.CAS
 
-import Numeric.Additive
 import Numeric.AffineSpace
 
 import Text.Read (readEither)
@@ -231,76 +218,6 @@ slowEpochGuard (ParentHeader p)
     | Mainnet01 <- _chainwebVersion p = _blockHeight p < 80000
     | otherwise = False
 {-# INLINE slowEpochGuard #-}
-
--- -------------------------------------------------------------------------- --
--- | BlockHeight
---
-newtype BlockHeight = BlockHeight { _height :: Word64 }
-    deriving (Eq, Ord, Generic)
-    deriving anyclass (NFData)
-    deriving newtype
-        ( Hashable, ToJSON, FromJSON
-        , AdditiveSemigroup, AdditiveAbelianSemigroup, AdditiveMonoid
-        , Num, Integral, Real, Enum
-        )
-instance Show BlockHeight where show (BlockHeight b) = show b
-
-instance IsMerkleLogEntry ChainwebHashTag BlockHeight where
-    type Tag BlockHeight = 'BlockHeightTag
-    toMerkleNode = encodeMerkleInputNode encodeBlockHeight
-    fromMerkleNode = decodeMerkleInputNode decodeBlockHeight
-    {-# INLINE toMerkleNode #-}
-    {-# INLINE fromMerkleNode #-}
-
-encodeBlockHeight :: MonadPut m => BlockHeight -> m ()
-encodeBlockHeight (BlockHeight h) = putWord64le h
-
-decodeBlockHeight :: MonadGet m => m BlockHeight
-decodeBlockHeight = BlockHeight <$> getWord64le
-
-encodeBlockHeightBe :: MonadPut m => BlockHeight -> m ()
-encodeBlockHeightBe (BlockHeight r) = putWord64be r
-
-decodeBlockHeightBe :: MonadGet m => m BlockHeight
-decodeBlockHeightBe = BlockHeight <$> getWord64be
-
--- -------------------------------------------------------------------------- --
--- Block Weight
---
--- This is the accumulated Hash difficulty
---
-newtype BlockWeight = BlockWeight HashDifficulty
-    deriving (Show, Eq, Ord, Generic)
-    deriving anyclass (NFData)
-    deriving newtype
-        ( Hashable
-        , ToJSON, FromJSON, ToJSONKey, FromJSONKey
-        , AdditiveSemigroup, AdditiveAbelianSemigroup
-        , Num
-        )
-
-instance IsMerkleLogEntry ChainwebHashTag BlockWeight where
-    type Tag BlockWeight = 'BlockWeightTag
-    toMerkleNode = encodeMerkleInputNode encodeBlockWeight
-    fromMerkleNode = decodeMerkleInputNode decodeBlockWeight
-    {-# INLINE toMerkleNode #-}
-    {-# INLINE fromMerkleNode #-}
-
-encodeBlockWeight :: MonadPut m => BlockWeight -> m ()
-encodeBlockWeight (BlockWeight w) = encodeHashDifficulty w
-{-# INLINE encodeBlockWeight #-}
-
-decodeBlockWeight :: MonadGet m => m BlockWeight
-decodeBlockWeight = BlockWeight <$> decodeHashDifficulty
-{-# INLINE decodeBlockWeight #-}
-
-encodeBlockWeightBe :: MonadPut m => BlockWeight -> m ()
-encodeBlockWeightBe (BlockWeight w) = encodeHashDifficultyBe w
-{-# INLINE encodeBlockWeightBe #-}
-
-decodeBlockWeightBe :: MonadGet m => m BlockWeight
-decodeBlockWeightBe = BlockWeight <$> decodeHashDifficultyBe
-{-# INLINE decodeBlockWeightBe #-}
 
 -- -------------------------------------------------------------------------- --
 -- Nonce
@@ -406,6 +323,10 @@ isLastInEpoch h = case effectiveWindow h of
 -- assume that a large amount of hash power has suddenly dropped out of the
 -- network. Thus we must perform Emergency Difficulty Adjustment to avoid
 -- stalling the chain.
+--
+-- NOTE: emergency DAs are now regarded a misfeature and have been disabled in
+-- all chainweb version. Emergency DAs are enabled (and have occured) only on
+-- mainnet01 for cut heights smaller than 80,000.
 --
 slowEpoch :: BlockHeader -> BlockCreationTime -> Bool
 slowEpoch p (BlockCreationTime ct) = actual > (expected * 5)
