+++ src/Chainweb/Mempool/InMem.hs
@@ -61,7 +61,7 @@ import System.Random
 -- internal imports
 
 import Chainweb.BlockHash
-import Chainweb.BlockHeader
+import Chainweb.BlockHeight
 import Chainweb.Logger
 import Chainweb.Mempool.InMemTypes
 import Chainweb.Mempool.Mempool
