+++ src/Chainweb/Miner/Coordinator.hs
@@ -67,6 +67,7 @@ import System.LogLevel (LogLevel(..))
 import Chainweb.BlockHash (BlockHash, BlockHashRecord(..))
 import Chainweb.BlockHeader
 import Chainweb.BlockHeader.Validation (prop_block_pow)
+import Chainweb.BlockHeight
 import Chainweb.Cut
 import Chainweb.Cut.CutHashes
 import Chainweb.CutDB
