+++ src/Chainweb/Pact/Backend/RelationalCheckpointer.hs
@@ -45,7 +45,7 @@ import Pact.Types.SQLite
 
 -- chainweb
 import Chainweb.BlockHash
-import Chainweb.BlockHeader
+import Chainweb.BlockHeight
 import Chainweb.Pact.Backend.ChainwebPactDb
 import Chainweb.Pact.Backend.Types
 import Chainweb.Pact.Backend.Utils
