+++ src/Chainweb/SPV/CreateProof.hs
@@ -42,6 +42,7 @@ import Prelude hiding (lookup)
 
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
+import Chainweb.BlockHeight
 import Chainweb.ChainId
 import Chainweb.Crypto.MerkleLog
 import Chainweb.CutDB
