+++ src/Chainweb/WebPactExecutionService/Types.hs
@@ -13,6 +13,7 @@ import Pact.Types.Hash
 
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
+import Chainweb.BlockHeight
 import Chainweb.ChainId
 import Chainweb.Mempool.Mempool (InsertError)
 import Chainweb.Miner.Pact
