+++ test/Chainweb/Test/Pact/ChainData.hs
@@ -38,6 +38,7 @@ import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeader.Genesis
 import Chainweb.BlockHeaderDB hiding (withBlockHeaderDb)
+import Chainweb.BlockHeight
 import Chainweb.ChainId
 import Chainweb.Difficulty
 import Chainweb.Mempool.Mempool (MempoolPreBlockCheck)
