+++ test/Chainweb/Test/Pact/SPV.hs
@@ -69,6 +69,7 @@ import Pact.Types.Term
 
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
+import Chainweb.BlockHeight
 import Chainweb.ChainId
 import Chainweb.Cut
 import Chainweb.CutDB
