+++ test/Chainweb/Test/Pact/TTL.hs
@@ -36,6 +36,7 @@ import Pact.Types.ChainMeta
 
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
+import Chainweb.BlockHeight
 import Chainweb.BlockHeader.Genesis
 import Chainweb.BlockHeaderDB hiding (withBlockHeaderDb)
 import Chainweb.Difficulty
