+++ test/Chainweb/Test/Pact/Utils.hs
@@ -130,6 +130,7 @@ import Pact.Types.Util (toB16Text)
 import Chainweb.BlockHeader
 import Chainweb.BlockHeader.Genesis
 import Chainweb.BlockHeaderDB hiding (withBlockHeaderDb)
+import Chainweb.BlockHeight
 import Chainweb.ChainId
 import Chainweb.Logger
 import Chainweb.Miner.Pact
